Move extensions/pkg/util/deployments.go to pkg/client/kubernetes/deployments.go
The duplicate extensions/pkg/util.ScaleDeployment function was deleted in favor of pkg/client/kubernetes.ScaleDeployment.
